do not add rm to file if doing a compile

+++ src/upsutl.c
@@ -61,6 +61,7 @@
 extern int g_LOCAL_VARS_DEF;
 extern char *g_temp_file_name;
 extern int g_keep_temp_file;
+extern int g_COMPILE_FLAG;
 
 /*
  * Declaration of private functions.
@@ -112,7 +113,7 @@ void upsutl_finish_temp_file( const FILE * const a_stream,
 
   /* we usually tell the file to delete itself.  however the user may
      override this */
-  if (! g_keep_temp_file) {
+  if (! g_COMPILE_FLAG && ! g_keep_temp_file) {
     fprintf((FILE *)a_stream, "/bin/rm -f %s\n", g_temp_file_name);
   }
 }
